move string
